correction actualise position + theta

+++ asser.h
@@ -2,4 +2,4 @@ void asser();
 void valide(int * reponse,int reponse_preced);
 int asser_done(int erreur_delta, int erreur_alpha);
 int arret_ok(int commande_moteur_D, int commande_moteur_G);
-float convert2PWM(int commande);
+int convert2PWM(int commande);
